Build number 02.00.02

+++ include/sharp/ir_types.h
@@ -1,6 +1,6 @@
 /* include/sharp/ir_types.h (sharp IrDA driver)
  *
- * Copyright (C) 2011 - 2014 SHARP CORPORATION All rights reserved.
+ * Copyright (C) 2011 - 2015 SHARP CORPORATION All rights reserved.
  *
  * This software is licensed under the terms of the GNU General Public
  * License version 2, as published by the Free Software Foundation, and
@@ -16,11 +16,11 @@
 #ifndef _IR_TYPES_H
 #define _IR_TYPES_H
 
-typedef unsigned long	uint32;
+typedef unsigned int	uint32;
 typedef unsigned short	uint16;
 typedef unsigned char	uint8;
 
-typedef long		int32;
+typedef int		int32;
 typedef short		int16;
 typedef char		int8;
